Refactor Variant class and associated genotype utility functions
