+++ python/tests/test_util.py
@@ -540,75 +540,3 @@ class TestRandomNuceotides:
     def test_length_bad_type(self, length):
         with pytest.raises(TypeError, match="argument must be a string"):
             tskit.random_nucleotides(length)
-
-
-#
-# Tests for allele_remap
-#
-@pytest.mark.parametrize(
-    "alleles_from, alleles_to, allele_map",
-    [
-        # Case 1: alleles_to is longer than alleles_from.
-        (
-            ["A", "C", "G", "T"],
-            ["G", "C"],
-            np.array([2, 1, 0, 3], dtype="uint32"),
-        ),
-        # Case 2: alleles_to is shorter than alleles_from.
-        (
-            ["G", "C"],
-            ["A", "C", "G", "T"],
-            np.array([2, 1], dtype="uint32"),
-        ),
-        # Case 3: alleles_to is empty.
-        (
-            ["A", "C", "G", "T"],
-            [],
-            np.array([0, 1, 2, 3], dtype="uint32"),
-        ),
-        # Case 4: alleles_from is empty.
-        (
-            [],
-            ["A", "C", "G", "T"],
-            np.array([], dtype="uint32"),
-        ),
-        # Case 5: Both lists are empty.
-        (
-            [],
-            [],
-            np.array([], dtype="uint32"),
-        ),
-        # Case 6: Both lists are tuples.
-        (
-            ("G", "C"),
-            ("A", "C", "G", "T"),
-            np.array([2, 1], dtype="uint32"),
-        ),
-        # Case 7: Both lists are numpy arrays.
-        (
-            np.array(("G", "C")),
-            np.array(("A", "C", "G", "T")),
-            np.array([2, 1], dtype="uint32"),
-        ),
-        # Case 8: Lists are of two different types.
-        (
-            np.array(("G", "C")),
-            ["A", "C", "G", "T"],
-            np.array([2, 1], dtype="uint32"),
-        ),
-        # Case 9: Lists contain elements of arbitrary types.
-        (
-            ["ABBA", "CDCD"],
-            ["ABBA", "CDCD", "EFEF", "GG", 18],
-            np.array([0, 1], dtype="uint32"),
-        ),
-        # Case 10: Lists contain unicode characters.
-        (
-            ["\u1F1E8", "\u1F1EC"],
-            ["\u1F1EC", "\u1F1E8", "\u1F1E6", "\u1F1F3"],
-            np.array([1, 0], dtype="uint32"),
-        ),
-    ],
-)
-def test_allele_remap(alleles_from, alleles_to, allele_map):
-    assert np.array_equal(allele_map, tskit.allele_remap(alleles_from, alleles_to))
